allow explicit naming of page files
Allow a property 'href' on a page to contain the name for the page. This
allows pages to link to each other. If not present, the default naming
scheme will be used (e0001.html etc).

If the property is present, it is the user's responsibility to make
certain this href does not contain a directory and to ensure uniqueness.

Note that this change allows for mixed content (i.e. default naming
scheme and explicit name intermingled).